amp/requirements.txt
# Core AI and Agent Framework
crewai>=0.11.0
langchain>=0.1.0

# Natural Language Processing
spacy>=3.7.0
textblob>=0.17.1

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.1
scikit-learn>=1.3.0
pydantic>=2.0.0

# Web Framework and Real-time Communication
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# HTTP Clients and API Integration
aiohttp>=3.8.0
requests>=2.31.0

# System Monitoring and Performance
psutil>=5.9.0

# Database Support
sqlite3  # Built into Python

# Environment Configuration
python-dotenv>=1.0.0

# XML Processing (for RSS feeds)
lxml>=4.9.0

# Optional: Enhanced web features
gunicorn>=21.2.0  # Production WSGI server
eventlet>=0.33.0  # SocketIO backend
redis>=4.5.0      # Optional: for production message queuing

# Development and Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Additional ML/NLP Libraries (optional but recommended)
nltk>=3.8.0
transformers>=4.30.0  # For advanced NLP models
torch>=2.0.0         # PyTorch for deep learning models

# Visualization and Reporting (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Security and Encryption (for production)
cryptography>=41.0.0
bcrypt>=4.0.0

# Logging and Structured Data
structlog>=23.1.0

# Time and Date Processing
python-dateutil>=2.8.0

# File Format Support
openpyxl>=3.1.0  # Excel files
python-docx>=0.8.11  # Word documents

# Additional API Connectors
tweepy>=4.14.0   # Twitter API alternative
praw>=7.7.0      # Reddit API wrapper

# Performance and Caching
cachetools>=5.3.0
joblib>=1.3.0
